+++ render/sdl/sdl_events.mojo
@@ -60,8 +60,8 @@ struct EventType(Indexer, Intable, TrivialRegisterPassable):
     var value: UInt32
 
     @always_inline
-    fn __init__(out self, value: Int):
-        self.value = UInt32(value)
+    fn __init__(out self, value: UInt32):
+        self.value = value
 
     @always_inline
     fn __int__(self) -> Int:
@@ -1381,7 +1381,7 @@ struct EventAction(Indexer, Intable, TrivialRegisterPassable):
     var value: UInt32
 
     @always_inline
-    fn __init__(out self, value: Int):
+    fn __init__(out self, value: UInt32):
         self.value = value
 
     @always_inline
